Main Menu functionality
Added a basic play and exit button, along with the code to switch from the 'main menu' scene to the 'player' scene when the 'play' button is pressed. As well as exit the game completely when the 'exit' button is pressed. Also added a way to effectively pause the game when the escape button is pressed, which prompts the main menu once again.